windows 32bit support
This commit changes a lot of things to make this npm module work on
Windows.

First, it requires that the compilation is done via the mingw64
compiler instead of the MSVC that is used by node-gyp. To do this the
ext2fs target is either a normal gyp target or a custom action depending
on a condition. The custom action invokes mingw32-make to do the
compilation of the library with GCC.

Second, the C++ source no longer includes the ext2fs headers from the
library. The reason is that they are too complex and bring in a lot of
dependencies that are unavailable when compiling the native node parts
in MSVC. To overcome this, we define a minimal header (src/ext2fs.h)
that defines only the structs/functions used by this module.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>