Makefile: disable NODEJS_CATCH_EXIT
Emscripten by default catches and handles exit(), which results in a
dump of the module, and necessitates a lot of scrolling to see the
actual Node error and relevant backtrace.

Since node-ext2fs is a library, disable this global error handler in
favor of letting the application do it.

See also:
https://github.com/emscripten-core/emscripten/blob/1.38.37/src/settings.js#L517-L524
https://github.com/kripken/sql.js/commit/b3d2c37f05bf60457025b3e04994f5a253f5092e

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>